% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_censorship}
\alias{check_censorship}
\title{Check if completely left or right censored}
\usage{
check_censorship(y)
}
\arguments{
\item{y}{The input feature matrix.}
}
\description{
Check if the target matrix is either completely left censored or completely 
right censored and print an error message accordingly.
}
