% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_surv_censorship}
\alias{check_surv_censorship}
\title{Check if completely left or right censored for a Surv object}
\usage{
check_surv_censorship(status)
}
\arguments{
\item{status}{The input status for Surv objects.}
}
\description{
Check if the target matrix is either completely left censored or completely 
right censored and print an error message accordingly for a Surv target matrix.
}
