% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.cv.lasso.R
\name{coef.cv.lasso}
\alias{coef.cv.lasso}
\title{Optimal coefficients for lasso, relaxed lasso or adaptive lasso fit of right-censored AFT model}
\usage{
\method{coef}{cv.lasso}(fit)
}
\arguments{
\item{fit}{object returned by \code{\link{cv.lasso}}.}
}
\value{
a named vector of length \code{nvar+1} with the optimal regression coefficients
}
\description{
This is a simple function to extract the optimal solution as determined by cross-validation from a
lasso, relaxed lasso or adaptive lasso fit.
}
\details{
The function returns the optimal regression coefficients for lasso, relaxed lasso or 
adaptive lasso fit of a right-censored AFT model as determined by cross-validation.
}
\examples{
library(survival)
X <- cbind(ovarian$ecog.ps, ovarian$rx)
y <- Surv(ovarian$futime, ovarian$fustat)
fit <- lasso(x=X, y=y, family="weibull")
cv.fit <- cv.lasso(fit)
coef.cv.lasso(cv.fit)

}
\seealso{
\code{\link{cv.lasso}}, \code{\link{scale.cv.lasso}}
}
\author{
Georg Heinze
}
