% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.iregnet.R
\name{coef.iregnet}
\alias{coef.iregnet}
\title{Extract model coefficients from iregnet object}
\usage{
\method{coef}{iregnet}(object, ...)
}
\arguments{
\item{object}{The S3 object of type \code{iregnet} returned by the
\code{iregnet} method.}

\item{...}{Optional arguments. Currently unused.}
}
\description{
Returns the model coefficients for the fit \link{iregnet} object. One model
is fit for each lambda value in \code{object$lambda}. Each of these is
returned in a single matrix of size \code{nvars} * \code{nobs}.
}

