% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.iregnet.R
\name{compute.loglik}
\alias{compute.loglik}
\title{Compute log likelihood}
\usage{
compute.loglik(y.mat, pred.mean, pred.scale, family)
}
\arguments{
\item{y.mat}{numeric matrix of output variables (n x 2)}

\item{pred.mean}{numeric matrix of mean parameters (predicted
values, n x p)}

\item{pred.scale}{numeric matrix of scale parameters (n x p)}

\item{family}{gaussian, logistic, weibull}
}
\value{
numeric matrix of log-likelihood values (n x p)
}
\description{
Compute log-likelihood of iregnet predictions, used to compute the
surrogate loss on the validation set in cross-validation
(cv.iregnet).
Note that for family="weibull" it will compute the likelihood only as sum, not as components
}
\author{
Toby Dylan Hocking
}
