% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_densities}
\alias{compute_densities}
\title{Compute the densities of given vector}
\usage{
compute_densities(z, j, family)
}
\arguments{
\item{z}{Vector of points at which the densities are to be calculated.}

\item{j}{If \code{1}, the returned vector contains
0, f, f'/f and f''/f, where, f is the
density function of the given distribution.
If 2, returned vector contains F, 1-F, f and f', where F is the
distribution function.}

\item{family}{The distribution of the data. \code{Guassian}, \code{logistic} and
(least) \code{extreme value} distributions are supported.}
}
\description{
C++ function to compute the density or distibution at each point of a
given vector. The supported distributions are \code{gaussian},
\code{logistic} and (least) \code{extreme value}.
\emph{This is used for testing, and may not be useful otherwise.}
}
