% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.lasso.R
\name{cv.lasso}
\alias{cv.lasso}
\title{Cross-validate lasso, relaxed lasso or adaptive lasso for right-censored AFT model}
\usage{
cv.lasso(fit, nfolds = 10, foldid = NULL, rep = 1, ...)
}
\arguments{
\item{fit}{object returned by \code{\link{lasso}}, \code{\link{relax.lasso}} or \code{\link{ada.lasso}}.}

\item{nfolds}{the number of folds for cross-validation (default=10)}

\item{foldid}{optional: assigned fold id for each observation. If supplied, overrides \code{nfolds}. It is
expected that \code{foldid} contains integer numbers of 1 to \code{nfolds}.}

\item{rep}{number of repetitions of cross-validation (default=1)}

\item{...}{Further arguments passed to \code{\link{iregnet}}}
}
\value{
Returns an object  with the following elements:\cr
\tabular{ll}{
 \code{call} \tab the call \cr
 \code{lasso.fit} \tab input fit object \cr
 \code{cvm} \tab cross validated log likelihood at each \code{lambda}.  \cr
 \code{cvm.sd} \tab standard deviation of fold-specific cross-validated log likelihood at each \code{lambda}. \cr
 \code{s} \tab index of optimal \code{lambda} \cr
 \code{lambda.min} \tab optimal \code{lambda} (maximizes CV log likelihood) \cr
 \code{lambda} \tab Vector of size \code{num_lambda} of (calculated or
  supplied) regularization parameter \code{lambda} values. \cr
 \code{omega} \tab The input \code{omega} \cr
 \code{x} \tab The input \code{x} matrix \cr
 \code{y} \tab The input \code{y} matrix (\code{\link{Surv}} object)\cr
 \code{type} \tab The type of estimation (taken from \code{fit$type}) \cr
 \code{family} \tab The input distribution (taken from \code{fit$family}) \cr
}
}
\description{
This is a simple function to crossvalidate a lasso,
relaxed lasso or adaptive lasso fit.
}
\details{
The function cross-validates a lasso, relaxed lasso or adaptive lasso to determine the optiomal 
value of $\lambda$. Repeated cross-validation is supported. The log likelihood is computed using a null fit from 
 \code{\link{survival::survreg}} and hence is already normalized to the scale of time variable.
}
\examples{
k<-10
n<-300
beta <- seq(0,1,1/(k-1))
X <- matrix(rnorm(k*n), n, k)
failtime <- rexp(n, 1/exp(10 + X \%*\% beta))
maxfu <- quantile(failtime, 0.5)
futime <- runif(n, 0, maxfu)
status <- (failtime < futime)*1
time <- pmin(failtime, futime)
y <- Surv(time, status)
fit<-lasso(y=y, x=X, family="weibull")
cvfit <- cv.lasso(fit)
plot.cv.lasso(cvfit)

}
\seealso{
\code{\link{relax.lasso}}, \code{\link{ada.lasso}}, \code{\link{lasso}}
}
\author{
Georg Heinze
}
