% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_cpp}
\alias{fit_cpp}
\title{C++ function to fit interval censored models with elastic net
  penalty.}
\usage{
fit_cpp(X, y, family, lambda_path, debug, out_status, intercept, alpha,
  scale_init, estimate_scale, unreg_sol, flag_standardize_x, max_iter,
  threshold, num_lambda, eps_lambda)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Output matrix in a 2 column format. \code{NA}s denote censoring.}

\item{family}{String denoting the distribution to be fit.}

\item{lambda_path}{Vector containing the path of hyper-parameter lambda
values.}

\item{debug}{Integer used for debugging during development. Unused otherwise.}

\item{out_status}{Vector containing censoring status of each observation.
If not provided, it will be calculated using the \code{y} matrix.}

\item{intercept}{If \code{true}, intercept is to be fit. The first column
of X must be \code{1}s if \code{intercept} is \code{true}.}

\item{alpha}{Hyper parameter for the elastic-net penalty.}

\item{scale_init}{The initial value of \code{scale} to be used for the fit.
If not provided, a value is calculated depending on the distribution.}

\item{estimate_scale}{If \code{true}, \code{scale} is estimated. Else, the
\code{scale} remains fixed at \code{scale_init}.}

\item{unreg_sol}{If \code{true}, the last model fit will be unregularized,
i.e., the final \code{lambda} value will be \code{0}. Only used if
\code{lambda_path} is not specified.}

\item{flag_standardize_x}{If \code{true}, the design matrix \code{X} will
column-wise standardized.}

\item{max_iter}{Maximum number of iterations to allow for \strong{each}
model (\code{lambda} value) along the regularization path.}

\item{threshold}{Convergence detected if absolute change in a coefficient
is less than this value.}

\item{num_lambda}{Number of lambda values to use in the regularization
path. Only used if \code{lambda_path} is not specified.}

\item{eps_lambda}{Ratio between the maximum and minimum values of
\code{lambda}. Maximum value of \code{lambda} is calculated based on
the distribution and the data. Only used if \code{lambda_path} is not
specified.}
}
\description{
\strong{This function is not meant to be called
on it's own!} Please use the \code{\link{iregnet}} function which includes
data validation and pre-processing.
}

