% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_status_from_surv}
\alias{get_status_from_surv}
\title{Get censoring status from a \code{Surv} object}
\usage{
get_status_from_surv(s)
}
\arguments{
\item{s}{\code{Surv} object.}
}
\description{
Returns the status column from a \link{Surv} object, and converts to the
form: right censored - 0, none - 1, left - 2, interval - 3.
}

