% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iregnet_compute_gradients}
\alias{iregnet_compute_gradients}
\title{Compute the gradients of the log Likelihood wrt beta, and scale}
\usage{
iregnet_compute_gradients(X, y, eta, scale, family)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Output matrix in 2 column format.}

\item{eta}{Linear predictors.}

\item{scale}{Scale.}

\item{family}{The distribution of the data. \code{Guassian}, \code{logistic} and
(least) \code{extreme value} distributions are supported.}
}
\description{
C++ function to compute the gradients of log likelihood with respect to the
coefficients \code{beta} and \code{scale}.
The supported distributions are \code{gaussian},
\code{logistic} and (least) \code{extreme value}.
\emph{This is used for testing, and may not be useful otherwise.}
}
