% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.iregnet.R
\name{logLik.iregnet}
\alias{logLik.iregnet}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{iregnet}(object, ...)
}
\arguments{
\item{object}{The result of an iregnet fit.}

\item{...}{Optional parameters, \emph{unused}.}
}
\description{
Extracts the Log-Likelihood from an object of class "iregnet" which has
already been fit.
}

