% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{plot.cv.iregnet}
\alias{plot.cv.iregnet}
\title{Regularization path plot for cross-validated iregnet model}
\usage{
\method{plot}{cv.iregnet}(x, ...)
}
\arguments{
\item{x}{object of class "cv.iregnet"}

\item{...}{ignored}
}
\value{
object of class "ggplot"
}
\description{
Plot the variable weights and estimated train/validation loss as a
function of the model complexity.
}
\author{
Toby Dylan Hocking
}

