% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.lasso.R
\name{plot.cv.lasso}
\alias{plot.cv.lasso}
\title{Plot cross-validated log likelihood vs. lambda for right-censored AFT model}
\usage{
\method{plot}{cv.lasso}(fit, adj = 5)
}
\arguments{
\item{fit}{object returned by \code{\link{cv.lasso}}.}

\item{adj}{adjustment value to change the spacing between nonzero values of \code{lambda} and 0 (default=5).}
}
\value{
no return value
}
\description{
This is a simple plot function to visualize the cross-validation results for a lasso,
relaxed lasso or adaptive lasso fit.
}
\details{
The function plots cross-validated log likelihood and standard errors against the penalty parameter.
The optimal solution is indicated by a horizontal and a vertical line, and a filled dot. The upper x-axis indicates
the number of active predictors.
}
\examples{
k<-10
n<-300
beta <- seq(0,1,1/(k-1))
X <- matrix(rnorm(k*n), n, k)
failtime <- rexp(n, 1/exp(10 + X \%*\% beta))
maxfu <- quantile(failtime, 0.5)
futime <- runif(n, 0, maxfu)
status <- (failtime < futime)*1
time <- pmin(failtime, futime)
y <- Surv(time, status)
fit<-lasso(y=y, x=X, family="weibull")
cvfit <- cv.lasso(fit)
plot.cv.lasso(cvfit)

}
\seealso{
\code{\link{cv.lasso}}
}
\author{
Georg Heinze
}
