% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.iregnet.R
\name{plot.iregnet}
\alias{plot.iregnet}
\title{Plot the path of the variables of iregnet fit
<TODO>}
\usage{
\method{plot}{iregnet}(x)
}
\description{
Plot the path of the variables of iregnet fit
<TODO>
}

