% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlasso.R
\name{plotlasso}
\alias{plotlasso}
\title{Plot lasso solution path for right-censored AFT model}
\usage{
plotlasso(fit, xvar = "log", intercept = TRUE)
}
\arguments{
\item{fit}{object returned by \code{\link{lasso}}, \code{\link{relax.lasso}} or \code{\link{ada.lasso}}.}

\item{xvar}{the meaning of the x-axis: either \code{"log"} (default; the logarithm of $\lambda$) 
or \code{"L1norm"}, the sum of absolute regression coefficients (except the intercept).}

\item{intercept}{if TRUE (default), adds the intercept path to the plot}
}
\value{
no return value
}
\description{
This is a simple plot function to visualize the solution path of a lasso,
relaxed lasso or adaptive lasso fit.
}
\details{
The function plots solution paths against the penalty parameter or the L1 norm of the regularized regression
coefficients. It uses standard colors to describe variables 1, 2, 3, ... A dashed line is used for the intercept.
}
\examples{
library(survival)
X <- cbind(ovarian$ecog.ps, ovarian$rx)
y <- Surv(ovarian$futime, ovarian$fustat)
fit <- lasso(x=X, y=y, family="weibull")
plotlasso(fit)

}
\seealso{
\code{\link{relax.lasso}}, \code{ada.lasso}, \code{\link{lasso}}
}
\author{
Georg Heinze
}
