% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.lasso.R
\name{predict.cv.lasso}
\alias{predict.cv.lasso}
\title{Predictions from lasso, relaxed lasso or adaptive lasso fit of right-censored AFT model}
\usage{
\method{predict}{cv.lasso}(fit, newdata, type = "response", ...)
}
\arguments{
\item{fit}{object returned by \code{\link{cv.lasso}}.}

\item{newdata}{a data set with the active predictors of the optimal solution.}

\item{type}{Type of prediction. See documentation of \code{\link{survival::predict.survreg}}. Default=\code{"response"}, which 
predicts expected time to failure.}
}
\value{
a vector with predicted values for each row of \code{newdata}
}
\description{
This is a simple function to predict from the optimal solution as determined by cross-validation of a
lasso, relaxed lasso or adaptive lasso fit.
}
\details{
The function predicts from the optimal solution as determined by cross-validation of a
lasso, relaxed lasso or adaptive lasso fit. It first generates a null fit from \code{survival::survreg} with the optimal solution
from \code{cv.lasso}. Hence, it inherits the functionality of \code{survival::predict.survreg}.
}
\examples{
library(survival)
X <- cbind(ovarian$ecog.ps, ovarian$rx)
y <- Surv(ovarian$futime, ovarian$fustat)
fit <- lasso(x=X, y=y, family="weibull")
cv.fit <- cv.lasso(fit)
predict.cv.lasso(cv.fit, newdata=ovarian)

}
\seealso{
\code{\link{cv.lasso}}, \code{\link{survival::predict.survreg}}
}
\author{
Georg Heinze
}
