\name{realNAN}
\alias{realNAN}
\docType{data}
\title{
  Data that caused a NAN error
}
\description{
  A penalty learning interval regression data set
  (H3K36me3_TDH_immune_FPOP test fold 1).
}
\usage{data("realNAN")}
\format{
  A named list with two elements: feature.mat (inputs) and target.mat (outputs).
}
\source{
  http://cbio.ensmp.fr/~thocking/data/penalty-learning-interval-regression-problems.tgz
}
\keyword{datasets}
