% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relax.lasso.R
\name{relax.lasso}
\alias{relax.lasso}
\title{Fit relaxed lasso for right-censored AFT model}
\usage{
relax.lasso(x, y, family, ...)
}
\arguments{
\item{x}{Input matrix of covariates with dimension n_obs * n_vars, with
\eqn{nvars \ge 2}. Sparse matrices are not supported.}

\item{y}{Response variable. Here, only \code{\link{Surv}} object is supported.}

\item{family}{The distribution to fit. It can be one of "gaussian", "logistic",
"loggaussian", "loglogistic", "extreme_value", "exponential", "weibull". Partial matching
is allowed.
\cr \emph{Default: "gaussian"}}

\item{...}{Further arguments passed to \code{\link{iregnet}}}
}
\value{
Returns an object  with the following elements:\cr
\tabular{ll}{
 \code{call} \tab the call \cr
 \code{coef} \tab Matrix of size \code{(n_vars+1) * num_lambda} containing
 intercept, coefficients of \code{X} for each \code{lambda} in the fit model.
   \cr
 \code{scale} \tab Vector of size \code{num_lambda} of estimated
   scale at each \code{lambda} value. \cr
 \code{lambda} \tab Vector of size \code{num_lambda} of (calculated or
  supplied) regularization parameter \code{lambda} values. \cr
 \code{x} \tab The input \code{x} matrix \cr
 \code{y} \tab The input \code{y} matrix (\code{\link{Surv}} object)\cr
 \code{type} \tab The type of estimation (\code{"relax.lasso"}) \cr
 \code{family} \tab The input distribution \cr
}
}
\description{
This is a simple wrapper function to fit accelerated failure time models 
using relaxed elastic net penalized maximum likeihood, calling \code{iregnet} and \code{survival::survreg}.
Supports gaussian, logistic, Weibull and extreme value distributions.
}
\details{
This is a wrapper function for \code{\link{iregnet}} with simpler
output, facilitating simple cross-validation. The function first calls
\code{\link{iregnet}}, and uses the active predictors at each \code{lambda}
(the sequence of \code{lambda} is determined by \code{\link{iregnet}})
to fit an AFT using \code{\link{survival::survreg}} for each \code{lambda}.
}
\examples{
library(survival)
k<-10
n<-300
beta <- c(rep(0, 4), seq(0.5, 2, length.out=6))
X <- matrix(rnorm(k*n), n, k)
failtime <- rexp(n, 1/exp(10 + X \%*\% beta))
maxfu <- quantile(failtime, 0.5)
futime <- runif(n, 0, maxfu)
status <- (failtime < futime)*1
time <- pmin(failtime, futime)
y <- Surv(time, status)
fit<-lasso(y=y, x=X, family="weibull")
plotlasso(fit, xvar="L1norm", intercept=FALSE)

}
\seealso{
\code{\link{lasso}}, \code{ada.lasso}, \code{\link{iregnet}}
}
\author{
Georg Heinze
}
