% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.cv.lasso.R
\name{scale.cv.lasso}
\alias{scale.cv.lasso}
\title{Optimal scale for lasso, relaxed lasso or adaptive lasso fit of right-censored AFT model}
\usage{
\method{scale}{cv.lasso}(fit)
}
\arguments{
\item{fit}{object returned by \code{\link{cv.lasso}}.}
}
\value{
numeric: the optimal scale parameter
}
\description{
This is a simple function to extract the optimal solution as determined by cross-validation from a
lasso, relaxed lasso or adaptive lasso fit.
}
\details{
The function returns the scale parameter at the optimal regression coefficients for 
lasso, relaxed lasso or adaptive lasso fit of a right-censored AFT model as determined by cross-validation.
}
\examples{
library(survival)
X <- cbind(ovarian$ecog.ps, ovarian$rx)
y <- Surv(ovarian$futime, ovarian$fustat)
fit <- lasso(x=X, y=y, family="weibull")
cv.fit <- cv.lasso(fit)
scale.cv.lasso(cv.fit)

}
\seealso{
\code{\link{cv.lasso}}, \code{\link{scale.cv.lasso}}
}
\author{
Georg Heinze
}
