% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stopifnot_error}
\alias{stopifnot_error}
\title{stopifnot with custom error message}
\usage{
stopifnot_error(err_message, ...)
}
\arguments{
\item{err_message}{The error message to print.}

\item{...}{An error is raised if any these expressions is \code{FALSE}.}
}
\description{
Like \code{stopifnot}, but with a custom error message.
}

