% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.cv.lasso.R
\name{surv.cv.lasso}
\alias{surv.cv.lasso}
\title{Survival predictions at specified time from lasso, relaxed lasso or adaptive lasso fit of right-censored AFT model}
\usage{
surv.cv.lasso(fit, newdata, time, grid = seq(0.001, 0.999, 0.001), se = TRUE)
}
\arguments{
\item{fit}{object returned by \code{\link{cv.lasso}}.}

\item{newdata}{a data set with the active predictors of the optimal solution.}

\item{time}{The time point for which survival should be predicted.}

\item{grid}{a grid of survival probabilities}

\item{se}{whether standard errors of survival probabilities should be returned (default=\code{TRUE})}
}
\value{
a vector with predicted survival probabilities for each row of \code{newdata}
}
\description{
This is a very simple function to predict survival from the optimal solution as determined by cross-validation of a
lasso, relaxed lasso or adaptive lasso fit at a specified time point.
}
\details{
The function uses a very simple method to determine survival probabilities from the optimal solution as determined by cross-validation of a
lasso, relaxed lasso or adaptive lasso fit at a specified time point. As \code{survival::predict.survreg} does not predict
survival probabilities but only quantiles, it first determines quantiles of the predicted survival curves for each row of \code{newdata}
and then looks up the survival probability which corresponds to a time just below the specified time point. The granularity of \code{grid} 
determines the precision of the survival probabilities.
}
\examples{
library(survival)
X <- cbind(ovarian$ecog.ps, ovarian$rx)
y <- Surv(ovarian$futime, ovarian$fustat)
fit <- lasso(x=X, y=y, family="weibull")
cv.fit <- cv.lasso(fit)
surv.cv.lasso(cv.fit, newdata=ovarian, time=)

}
\seealso{
\code{\link{cv.lasso}}, \code{\link{survival::predict.survreg}}
}
\author{
Georg Heinze
}
