% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydf.R
\name{tidydf}
\alias{tidydf}
\title{Return a tidy data.frame from iregnet fit}
\usage{
tidydf(x, ...)
}
\arguments{
\item{x}{The S3 object of type \code{iregnet} returned by the \code{iregnet}
method.}

\item{...}{Other parameters. Currently unused.}
}
\description{
Returns a tidy \link{data.frame} from a fitted "iregnet" object.
}
\details{
This function is used to obtain an intermediate \code{data.frame} used in
\link{plot.iregnet}.
It can be used for producing other plots using \code{ggplot2}.
NOTE: \code{Intercept} (if present) is \strong{not} included in the
\code{arclength} since it is never regularized.
}

