% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{transformed_distributions}
\alias{transformed_distributions}
\title{List of non-basic distributions in terms of basic distributions}
\format{An object of class \code{list} of length 4.}
\usage{
transformed_distributions
}
\description{
A \code{list} containing the supported distributions that are
transformations of other "base" distributions. For instance,
\code{loggaussian} is the same as \code{gaussian} after log transforming the
output \code{y}.
}
\details{
The list contains the transformation \code{trans}, the base distribution
\code{dist},  and the inverse transformation \code{itrans}. It is used
internally to fit such distributions.
}
\keyword{datasets}
