% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.iregnet.R
\name{cv.iregnet}
\alias{cv.iregnet}
\title{Cross-validation for iregnet}
\usage{
cv.iregnet(x, y, family, nfolds, foldid, ...)
}
\arguments{
\item{x}{numeric matrix of input features (n x p)}

\item{y}{numeric matrix of output variables (n x 2)}

\item{family}{gaussian, logistic}

\item{nfolds}{positive integer between 2 and n, by default 10.}

\item{foldid}{integer vector of length n (fold for each observation), by default we use nfolds.}

\item{...}{passed to iregnet for both the full and cv fits.}
}
\value{
model fit list of class "cv.iregnet"
}
\description{
First fit iregnet to the entire data set, then use K-fold
cross-validation to estimate the validation error of each lambda
parameter.
}
\author{
Toby Dylan Hocking
}
