% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_cpp}
\alias{fit_cpp}
\title{C++ function to fit regularized AFT models with interval censored data}
\usage{
fit_cpp(X, y, family, lambda_path, debug, out_status, intercept, alpha,
  scale_init, estimate_scale, unreg_sol, flag_standardize_x, max_iter,
  threshold, num_lambda, eps_lambda, thresh_divergence = 1)
}
\description{
\strong{NOTE:} This function is not meant to be called on it's own! Please use
the \code{\link{iregnet}} function.
}

