% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.iregnet.R
\name{plot.iregnet}
\alias{plot.iregnet}
\title{Plot the path of the variables of iregnet fit}
\usage{
\method{plot}{iregnet}(x, xvar = c("norm", "lambda"), label = TRUE, ...)
}
\arguments{
\item{xvar}{Variable on the X-axis against which the coefficients are
plotted. \code{norm} plots against the L1 norm of the coefficients, i.e.,
arclength. \code{lambda} plots against log-lambda sequence.}

\item{label}{If \code{TRUE}, coefficient names / variable sequence numbers
are plotted along with the curves.}

\item{...}{Optional parameters to be passed to the \code{matplot} function.}

\item{fit}{The S3 object of type \code{iregnet} returned by the \code{iregnet}
method.}
}
\description{
Produces a coefficient profile plot of the coefficient paths for a fitted
"iregnet" object.
}
\details{
\code{Intercept} (if present) is \strong{not} included in the \code{arclength}
since it is never regularized. It is also not plotted.
This function uses \link{tidymatrix} function to obtain a matrix from the
\link{iregnet} object. It can be directly used to produce other plots.
}

