% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.iregnet.R
\name{predict.cv.iregnet}
\alias{predict.cv.iregnet}
\title{predict CV iregnet}
\usage{
\method{predict}{cv.iregnet}(object, newx, type = "min", ...)
}
\arguments{
\item{object}{result of cv.iregnet}

\item{newx}{feature matrix}

\item{type}{min or 1sd}

\item{...}{further parameters}
}
\value{
predicted values
}
\description{
Predict CV iregnet
}
\details{
Get predicted values
}
\author{
Toby Dylan Hocking
}
