% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.iregnet.R
\name{print.iregnet}
\alias{print.iregnet}
\title{Print the iregnet fit}
\usage{
\method{print}{iregnet}(x, ...)
}
\arguments{
\item{x}{The result of an iregnet fit.}

\item{...}{Optional parameters for print. If \code{n} is supplied, only the
first \code{n} models will be printed.}
}
\description{
Prints a summary of the results of the
}
